+++ plugins/provisioners/docker/cap/debian/docker_configure_auto_start.rb
@@ -9,9 +9,13 @@ module VagrantPlugins
                 comm.sudo("echo 'DOCKER_OPTS=\"-r=true ${DOCKER_OPTS}\"' >> /etc/default/docker")
                 comm.sudo("stop docker")
                 comm.sudo("start docker")
+
+                # Wait some amount time for the pid to become available
+                # so that we don't start executing Docker commands until
+                # it is available.
                 [0, 1, 2, 4].each do |delay|
                   sleep delay
-                  break if comm.test('test -f /var/run/docker.pid')
+                  break if machine.guest.capability(:docker_daemon_running)
                 end
               end
             end